+++ magda-web-client/src/Components/FeedbackForm.js
@@ -10,6 +10,7 @@ import feedback from "../assets/feedback.svg";
 import close from "../assets/close.svg";
 import success from "../assets/success.svg";
 import Notification from "../UI/Notification";
+import ReactTooltip from 'react-tooltip'
 
 class FeedbackForm extends React.Component {
     constructor(props) {
@@ -18,15 +19,40 @@ class FeedbackForm extends React.Component {
             isOpen: false,
             name: "",
             email: "",
-            feedback: ""
+            feedback: "",
+            validationErrorMessage: this.checkRequiredFields(this.state)
         };
+
         this.onCancel = this.onCancel.bind(this);
         this.changeValue = this.changeValue.bind(this);
         this.onSubmit = this.onSubmit.bind(this);
         this.onDismissNotification = this.onDismissNotification.bind(this);
     }
 
+    /**
+     * Make sure different wording will be used depends on no. of the field items
+     */
+    createFieldListString(fields) {
+        if (!fields || !fields.length) return null;
+        if (fields.length === 1) return `${fields[0]} field is`;
+        const lastField = fields.pop();
+        return fields.join(", ") + ` & ${lastField} fields are`;
+    }
+
+    checkRequiredFields(state) {
+        const requiredFields = [];
+        if (!state || !state.email || state.email.trim() === '') requiredFields.push(`Email`);
+        if (!state || !state.feedback || state.feedback.trim() === '') requiredFields.push(`Feedback`);
+        if (requiredFields.length) return this.createFieldListString(requiredFields) + " mandatory.";
+        return null;
+    }
+
     onSubmit() {
+        /**
+         * It seems once the button is disabled, mouse related events won't trigger anymore --- this will break the tooltip.
+         * Instead of having button disabled, we will check here to make sure feedback is only sent when no validation error.
+         */
+        if (this.state.validationErrorMessage) return;
         this.props.fetchFeedback(
             JSON.stringify({
                 name: this.state.name,
@@ -41,13 +67,20 @@ class FeedbackForm extends React.Component {
             isOpen: false,
             name: "",
             email: "",
-            feedback: ""
+            feedback: "",
+            validationErrorMessage: this.checkRequiredFields()
         });
     }
 
     changeValue(key, event) {
-        this.setState({
+        let curState = {
+            ...this.state,
             [key]: event.target.value
+        };
+
+        this.setState({
+            [key]: event.target.value,
+            validationErrorMessage: this.checkRequiredFields(curState)
         });
     }
 
@@ -57,8 +90,15 @@ class FeedbackForm extends React.Component {
             isOpen: false,
             name: "",
             email: "",
-            feedback: ""
+            feedback: "",
+            validationErrorMessage: this.checkRequiredFields()
         });
+    }
+
+    onDismissErrorNotification() {
+        this.setState({
+            errorMessage: null
+        })
     }
 
     renderByState() {
@@ -110,12 +150,12 @@ class FeedbackForm extends React.Component {
                         onChange={this.changeValue.bind(this, "name")}
                     />
                     <Input
-                        label="Email"
+                        label="Email (*)"
                         value={this.state.email}
                         onChange={this.changeValue.bind(this, "email")}
                     />
                     <Textarea
-                        label="Feedback"
+                        label="Feedback (*)"
                         value={this.state.feedback}
                         onChange={this.changeValue.bind(this, "feedback")}
                     />
@@ -127,15 +167,33 @@ class FeedbackForm extends React.Component {
                         >
                             Cancel
                         </Button>
-                        <Button
-                            className="send-btn"
-                            disabled={this.props.isSendingFeedback}
-                            onClick={this.onSubmit}
-                        >
-                            {this.props.isSendingFeedback
-                                ? "Sending..."
-                                : "Send"}
-                        </Button>
+                        {this.state.validationErrorMessage ? (
+                            <Button
+                                className="send-btn disabled-looking-button"
+                                disabled={false}
+                                onClick={this.onSubmit}
+                                data-tip={this.state.validationErrorMessage}
+                                data-place="top"
+                            >
+                                {this.props.isSendingFeedback
+                                    ? "Sending..."
+                                    : "Send"}
+                            </Button>
+                        ) : (
+                                <Button
+                                    className="send-btn"
+                                    disabled={this.props.isSendingFeedback}
+                                    onClick={this.onSubmit}
+                                >
+                                    {this.props.isSendingFeedback
+                                        ? "Sending..."
+                                        : "Send"}
+                                </Button>
+                            )}
+
+                        {this.state.validationErrorMessage ? (
+                            <ReactTooltip />
+                        ) : null}
                     </div>
                 </div>
             </div>